chore: Moves machine code to its own file
